Modularity and Type hinting
