--- trackers/__init__.py
@@ -1,1 +1,0 @@
-from .tracker import Tracker
